CSN: CSN-hasim-models-22
Added separate dummy memory modules for ICache and DCache. Modified
direct-mapped and set-associative ICache to service requests from the
CPU despite processing a previous miss in the background. Added a
write-through direct mapped data cache that currently takes one input
port. A future modification will be the separation of speculative
input ports and commit input ports.
Modified unpipeline module to deal with new ICache in conjunction with
DCache.
Created a separate DCache types file to gather paramerization of the
DCache infrastructure.
Modified cache interface to take both the PC and the address
referenced for reads and writes to the data cache.