CSN: CSN-hasim-models-26
Added direct-mapped, write-through data cache code. Data cache has two
ports (one of which handles loads and the other commits
stores). During stall times, miss retry signals and invalids are sent
back to the pipeline as appropriate. It is assumed that a separate
store buffer module will be created to take care of the effects of
speculative stores and will deal with corresponding aliasing issues.