CSN: CSN-hasim-models-59
Added support for drainBefore/drainAfter to OOO model. It still runs out of
tokens, but this does remove the warnings from M5 about executing a halt
instruction.

Murali should look this over to see if the scheme is reasonable (it's basically
the same as the one I added to the inorder model).

Michael P.