CSN: CSN-hasim-models-72
Front end next PC logic assumed a full bundle of instructions is fetched
on each cycle and blindly took the next PC from the branch predictor.  When
a full bundle was not fetched instructions were skipped.  All I did was
ignore the branch predictor's idea of the next PC when a full bundle isn't
fetched.  The predictor interface is as broken as ever and IPC of the
modeled machine is still abysmal.  memory_tests_aligned now runs correctly,
though IPC is 0.09.