CSN: CSN-hasim-models-113
Fix a deadlock with multiple contexts. It seems that scratchpad readRsp/writeReq
should not be called in the same rule. By moving it to a different rule all
benchmarks pass with 2 contexts.

The reason this bug only manifested when running with 2 contexts is that 1
context would never fill up a FIFO2.

Michael P.