CSN: CSN-hasim-models-163
Move branch predictor storage to BRAM.  The LUTRAM storage required for a 64-way
model was quite large.  The new code maintains the same semantics.  Prediction
requests are locked out for a CPU while updates to that CPU are in progress.
A more refined address comparison could be done, but the typical stage controller
of a caller would have this same policy already.