CSN: CSN-hasim-models-165
The default mkStageController() provides only a 2-entry FIFO, which is fine
for a normal pipeline.  When a timing model uses both a stage controller and
makes a request to some other pipeline with a rendezvous in the next stage,
the stage controller buffer needs to be as large as the pipeline depth of
the external request.  This is particularly important for multi-FPGA
configurations where the latency may be quite high.

Models are now changed to use mkBufferedStageController() when an external
pipeline is embedded within a set of stages.