CSN: CSN-hasim-models-176
Fix model deadlock:  sending a request to the functional partition should be treated as
a channel with credits.  The caller must guarantee to provide buffering space to receive
all outstanding responses.  Failure to do so can lead to deadlocks in the functional
partition during contended operations, such as rewind.  In the most recent deadlock,
rewind backed up due to lack of buffering and a dependence chain in the timing model
that didn't allow rewind to proceed until some other functional operation could
complete.  The functional operation was blocked by the rewind.

Debugged completely on hardware using debug scan!