CSN: CSN-hasim-models-178
Rewind and get dependencies share a pipeline in the model but may conflict
in the functional model if rewind is attempted while too many dependency
requests are in flight (even for other contexts).  Limit pipelined requests
to either one category or the other, allowing at most one of a type if
the other type is already in flight.

This affects model simulation rate only.  Since rewinds are relatively
infrequent the model performance is about the same.