CSN: CSN-hasim-models-185
Changes to inorder model:

- Back to back write then read of a register caused a one cycle bubble.
  Change the target machine to add a bypass for this case by setting
  "Exe_to_Dec_writeback" port latency to 0.

- Move branch predictor and BTB to the head of the pipeline, replacing
  the line predictor.

- Change fetch to ITLB latency to 2 to simulate 2 cycle branch prediction.

In typical programs, removing the terrible line predictor (always PC+4)
and replacing it with a branch predictor has no effect on performance!
This was especially true before adding the bypass for register-to-register.
The change does cause fewer rewinds, but the 3 cycle load to use bubble
slows the machine and the instruction queue hides many front end address
prediction problems.