CSN: CSN-hasim-models-212
The OCN read interface is now rich enough to deal with a client that reads
an incoming message but choses not to dequeue it because the client doesn't
have anywhere to put the new message.  The previous code worked, mostly,
because the OCN is typically the rate limiting step and clients are
typically always ready to receive a message.

Both the memory and LLC clients now use the interface properly.