CSN: CSN-hasim-models-213
Instead of sending a real packet across the simulated network, we send narrow
messages with routing information and a pointer to the full packet.  This
saves a lot of FPGA memory since the network FIFOs can be smaller.  Until
now, the only messages were loads and stores.  In order to support coherence
traffic and requests to remove LLC segments the messages types must be
richer.  This is the first step.

icn-packet-payload.bsv defines a service in which packet descriptors are
stored.  In the current code, only the LLC actually uses the service.