CSN: CSN-hasim-models-231
Provide a base module for connecting a vector of stall ports to an on-chip
network.  The module maps vector slots to lanes in the network.  Saving
message data in a side buffer instead of encoding the entire message as
flits is handled internally.

The ports may have different data types, though all types must belong
to the OCN_SEND_TYPE typclass.