CSN: CSN-hasim-models-240
Cache protocol components now communicate with each other using CACHE_PROTOCOL_MSG
instead of MEMORY_REQ/MEMORY_RSP.  This was not strictly necessary in the
no-coherence model, but these will serve as the template for coherent caches.
They also share components, such as the L1 arbiter, that can now share a type
across coherence protocol families.

The no-coherence LLC still communicates using MEMORY_REQ/MEMORY_RSP.  This
is because it may be used in coherence frameworks where the LLC is a simple
distributed cache, outside of the coherent caches.